+++ test/stats/results_test.csv
@@ -1,35 +1,35 @@
 ,pattern, -1(top1),+1(top1),p-c-,p+c+,p-c+,p+c-,p-c=,p+c=
-0,Asserts,0,0,1,1,0,0,0,0
-1,Setters,0,0,1,1,0,0,0,0
-2,Empty Rethrow,0,0,1,1,0,0,0,0
-3,Prohibited class name,1,0,2,2,0,0,0,0
-4,Force Type Casting,7,0,12,12,0,0,0,0
-5,Count If Return,0,0,8,8,0,0,0,0
-6,Implements Multi,1,0,18,18,0,0,0,0
-7,Instance of,0,0,4,4,0,0,0,0
-8,Many primary constructors,8,0,18,18,0,0,0,0
-9,Method chain,0,0,17,17,0,0,0,0
-10,Multiple try,0,0,10,10,0,0,0,0
-11,Non final attribute,54,0,98,98,0,0,0,0
-12,Null check,5,0,45,45,0,0,0,0
-13,Partial synchronized,0,0,1,1,0,0,0,0
-14,Redundant catch,0,0,1,1,0,0,0,0
-15,Return null,0,0,11,11,0,0,0,0
-16,String concat,10,0,12,12,0,0,0,0
-17,Super Method,4,0,8,8,0,0,0,0
-18,This in constructor,0,0,1,1,0,0,0,0
-19,Var declaration distance for 5 lines,0,0,8,8,0,0,0,0
-20,Var declaration distance for 7 lines,0,0,7,7,0,0,0,0
-21,Var declaration distance for 11 lines,0,0,6,6,0,0,0,0
-22,Var in the middle,10,0,24,24,0,0,0,0
-23,Array as function argument,0,0,2,2,0,0,0,0
-24,Joined validation,0,0,2,2,0,0,0,0
-25,Non final class,36,0,140,140,0,0,0,0
-26,Private static method,0,0,3,3,0,0,0,0
-27,Public static method,0,0,5,5,0,0,0,0
-28,Var siblings,0,0,6,6,0,0,0,0
-29,Null Assignment,0,0,7,7,0,0,0,0
-30,Multiple While,0,0,8,8,0,0,0,0
-31,Protected Method,0,0,2,2,0,0,0,0
-32,Send Null,1,0,2,2,0,0,0,0
-33,Nested Loop,3,0,8,8,0,0,0,0
+0,Asserts,2,6,2,2,7,7,0,0
+1,Setters,0,0,2,2,8,8,0,0
+2,Empty Rethrow,0,0,2,2,8,8,0,0
+3,Prohibited class name,0,0,2,2,8,8,0,0
+4,Force Type Casting,0,0,2,2,8,8,0,0
+5,Count If Return,0,0,2,2,8,8,0,0
+6,Implements Multi,0,0,2,2,8,8,0,0
+7,Instance of,0,0,2,2,8,8,0,0
+8,Many primary constructors,0,0,2,2,8,8,0,0
+9,Method chain,0,0,2,2,8,8,0,0
+10,Multiple try,0,0,2,2,8,8,0,0
+11,Non final attribute,0,0,2,2,8,8,0,0
+12,Null check,0,0,2,2,8,8,0,0
+13,Partial synchronized,0,0,2,2,8,8,0,0
+14,Redundant catch,0,0,2,2,8,8,0,0
+15,Return null,0,0,2,2,8,8,0,0
+16,String concat,0,1,2,2,8,8,0,0
+17,Super Method,0,0,2,2,8,8,0,0
+18,This in constructor,0,0,2,2,8,8,0,0
+19,Var declaration distance for 5 lines,0,0,2,2,8,8,0,0
+20,Var declaration distance for 7 lines,0,0,2,2,8,8,0,0
+21,Var declaration distance for 11 lines,0,0,2,2,8,8,0,0
+22,Var in the middle,0,0,2,2,8,8,0,0
+23,Array as function argument,0,0,2,2,8,8,0,0
+24,Joined validation,0,0,2,2,8,8,0,0
+25,Non final class,0,0,2,2,8,8,0,0
+26,Private static method,0,0,2,2,8,8,0,0
+27,Public static method,0,0,2,2,8,8,0,0
+28,Var siblings,0,0,2,2,8,8,0,0
+29,Null Assignment,0,0,2,2,8,8,0,0
+30,Multiple While,0,1,2,2,8,8,0,0
+31,Protected Method,0,0,2,2,8,8,0,0
+32,Send Null,0,0,2,2,8,8,0,0
+33,Nested Loop,0,0,2,2,8,8,0,0
